requirements.txt
apache-airflow==2.10.4
scikit-learn
pandas
numpy
matplotlib
joblib
